+++ fluxburst/kubernetes/plugins.py
@@ -22,10 +22,6 @@ class KubernetesBurstPlugin(BurstPlugin):
     An additional wrapper to the plugin that adds support for the Flux Operator
     """
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.clusters = {}
-
     def ensure_namespace(self, kubectl):
         """
         Use the instantiated kubectl to ensure the cluster namespace exists.
@@ -40,6 +36,38 @@ class KubernetesBurstPlugin(BurstPlugin):
             logger.warning(
                 f"🥵️ Issue creating namespace {self.params.namespace}, assuming already exists."
             )
+
+    def validate(self):
+        """
+        Validate ensures that required conditions are met.
+
+        This should return True/False to indicate if the plugin is valid or not.
+        """
+        # If we aren't doing an isolated burst, we require host names, configs, etc.
+        if not self.params.isolated_burst:
+            # Lead host and port are required. A custom broker.toml can be provided,
+            # but we are having the operator create it for us
+            if (
+                not self.params.lead_port
+                or not self.params.lead_host
+                or not self.params.lead_size
+            ):
+                logger.warning(
+                    "Parameteres lead_host, lead_size, and lead_port must be defined when not doing an isolated_burst."
+                )
+                return False
+
+            print(
+                f"Broker lead will be accessible on {self.params.lead_host}:{self.params.lead_port}"
+            )
+
+            # We also need to have the munge key and curve cert
+            if self.params.munge_key and not os.path.exists(self.params.munge_key):
+                logger.warning(
+                    f"Provided munge key {self.params.munge_key} does not exist."
+                )
+                return False
+        return True
 
     def check_configs(self):
         """
@@ -64,38 +92,60 @@ class KubernetesBurstPlugin(BurstPlugin):
                 f"Issue installing the operator: {exc}, assuming already exists"
             )
 
-    def run(self):
+    def run(self, request_burst=False, nodes=None, tasks=None):
         """
         Given some set of scheduled jobs, run bursting.
+
+        If request_burst is True, a specific number of tasks and size
+        must be provided to request a cluster in advance (that jobs can
+        be run on).
         """
         # Exit early if no jobs to burst
         if not self.jobs:
             logger.info(f"Plugin {self.name} has no jobs to burst.")
             return
 
+        # If we have requested a burst, tasks and size are required
+        if request_burst and (not tasks or not nodes):
+            logger.warning("Burst requests require nodes and tasks.")
+            return
+
         # Ensure we have a flux operator yaml file, fosho, foyaml!
         foyaml = helpers.ensure_flux_operator_yaml(self.params.flux_operator_yaml)
 
         # lead host / port / size / are required in the dataclass
+        # But ONLY if this isnt' an isolated burst!
         # We check munge paths here, because could be permissions issue
-        self.check_configs()
+        if not self.params.isolated_burst:
+            self.check_configs()
 
+        # This call registers the name of the cluster to self.clusters
         cli = self.create_cluster()
         kubectl = cli.get_k8s_client()
 
         # Install the operator!
         self.install_flux_operator(kubectl, foyaml)
 
+        # Are we requesting or running jobs?
+        if request_burst:
+            self.create_minicluster(kubectl, "sleep infinity", nodes, tasks)
+        else:
+            self.run_jobs(kubectl)
+
+    def run_jobs(self, kubectl):
+        """
+        Run jobs (creating MiniClusters), assuming that the burst has been done.
+        """
         # Create a MiniCluster for each job
         for _, job in self.jobs.items():
-            self.create_minicluster(kubectl, job)
+            command = " ".join(job["spec"]["tasks"][0]["command"])
+            self.create_minicluster(kubectl, command, job["nnodes"], job["ntasks"])
 
-    def create_minicluster(self, kubectl, job):
+    def create_minicluster(self, kubectl, command, nodes, tasks):
         """
         Create the MiniCluster
         """
-        command = " ".join(job["spec"]["tasks"][0]["command"])
-        logger.info(f"Preparing MiniCluster for {job['id']}: {command}")
+        logger.info(f"Preparing MiniCluster for {command}")
 
         # The plugin is assumed to be running from the lead broker
         # of the cluster it is bursting from, this we get info about it
@@ -111,8 +161,8 @@ class KubernetesBurstPlugin(BurstPlugin):
             cpu_limit=self.params.cpu_limit,
             namespace=self.params.namespace,
             broker_toml=self.params.broker_toml,
-            tasks=job["ntasks"],
-            size=job["nnodes"],
+            tasks=tasks,
+            size=nodes,
             image=self.params.image,
             wrap=self.params.wrap,
             log_level=self.params.log_level,
@@ -130,7 +180,9 @@ class KubernetesBurstPlugin(BurstPlugin):
         # Let's assume there could be bugs applying this differently
         crd_api = kubernetes_client.CustomObjectsApi(kubectl.api_client)
 
-        self.ensure_secrets(kubectl)
+        # These are not needed if we are doing an isolated burst
+        if not self.params.isolated_burst:
+            self.ensure_secrets(kubectl)
 
         # Create the MiniCluster! This also waits for it to be ready
         print(
@@ -139,7 +191,11 @@ class KubernetesBurstPlugin(BurstPlugin):
 
         # Make sure we provide the core_v1_api we've created
         operator = FluxMiniCluster(core_v1_api=kubectl)
-        return operator.create(**minicluster, container=container, crd_api=crd_api)
+        try:
+            return operator.create(**minicluster, container=container, crd_api=crd_api)
+        except Exception as e:
+            print(f"Issue creating cluster, does it already exist?: {e}")
+            return True
 
     def ensure_secrets(self, kubectl):
         """