+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/converged-computing/flux-burst/tree/main) (0.0.x)
+ - plugin-specific validation, ability to request burst (0.0.12)
  - support for mock mode, meaning that we provide a fake job (0.0.11)
  - shared logic in kubernetes module for gke and eks (0.0.1)
  - initial skeleton release of project (0.0.0)